refac: stat 부분 getChartOption 함수 추출
